ut(ios): ut for properties methods and constants for Peripheral manager
